recent and favorite related problem solved
When I visit a playlist or add a playlist to favorite, then if I delete this and go to favorite or recent page, an error occured because of undefined url. Successfully solved this problem by creating remove function to remove this playlist from favorite and recent when it would be removed.